modify/different commit func for update github repo
